+++ back/src/bsvhu/__tests__/factories.vhu.ts
@@ -9,31 +9,55 @@ import { prisma } from "@td/prisma";
 import {
   companyFactory,
   ecoOrganismeFactory,
-  siretify
+  siretify,
+  userWithCompanyFactory
 } from "../../__tests__/factories";
 import { BsvhuForElastic, BsvhuForElasticInclude } from "../elastic";
+import { getUserCompanies } from "../../users/database";
 
 export const bsvhuFactory = async ({
+  userId,
   opt = {}
 }: {
+  userId?: string;
   opt?: Partial<Prisma.BsvhuCreateInput>;
 }): Promise<BsvhuForElastic> => {
   const transporterCompany = await companyFactory({
     companyTypes: ["TRANSPORTER"]
   });
   const destinationCompany = await companyFactory({
-    companyTypes: ["WASTE_VEHICLES"]
+    companyTypes: ["WASTE_VEHICLES"],
+    wasteVehiclesTypes: ["BROYEUR", "DEMOLISSEUR"]
+  });
+  const brokerCompany = await companyFactory({
+    companyTypes: ["BROKER"]
+  });
+  const traderCompany = await companyFactory({
+    companyTypes: ["TRADER"]
   });
   const ecoOrganisme = await ecoOrganismeFactory({
     handle: { handleBsvhu: true },
     createAssociatedCompany: true
   });
+  const bsvhuObject = getVhuFormdata();
+  const userAndCompany = userId
+    ? null
+    : await userWithCompanyFactory("ADMIN", {
+        siret: bsvhuObject.emitterCompanySiret,
+        name: bsvhuObject.emitterCompanyName ?? undefined,
+        address: bsvhuObject.emitterCompanyAddress ?? undefined,
+        contact: bsvhuObject.emitterCompanyContact ?? undefined,
+        contactPhone: bsvhuObject.emitterCompanyPhone ?? undefined,
+        contactEmail: bsvhuObject.emitterCompanyMail ?? undefined
+      });
   const created = await prisma.bsvhu.create({
     data: {
-      ...getVhuFormdata(),
+      ...bsvhuObject,
       transporterCompanySiret: transporterCompany.siret,
       destinationCompanySiret: destinationCompany.siret,
       ecoOrganismeSiret: ecoOrganisme.siret,
+      brokerCompanySiret: brokerCompany.siret,
+      traderCompanySiret: traderCompany.siret,
       ...opt
     },
     include: {
@@ -45,9 +69,35 @@ export const bsvhuFactory = async ({
         .flatMap(intermediary => [intermediary.siret, intermediary.vatNumber])
         .filter(Boolean)
     : [];
+
+  // For drafts, only the owner's sirets that appear on the bsd have access
+  const canAccessDraftOrgIds: string[] = [];
+  if (created.isDraft) {
+    const userCompanies = await getUserCompanies(
+      (userId || userAndCompany?.user?.id) as string
+    );
+    const userOrgIds = userCompanies.map(company => company.orgId);
+    const bsvhuOrgIds = [
+      ...intermediariesOrgIds,
+      created.emitterCompanySiret,
+      ...[
+        created.transporterCompanySiret,
+        created.transporterCompanyVatNumber
+      ].filter(Boolean),
+      created.ecoOrganismeSiret,
+      created.destinationCompanySiret,
+      created.traderCompanySiret,
+      created.brokerCompanySiret
+    ].filter(Boolean);
+    const userOrgIdsInForm = userOrgIds.filter(orgId =>
+      bsvhuOrgIds.includes(orgId)
+    );
+    canAccessDraftOrgIds.push(...userOrgIdsInForm);
+  }
   return prisma.bsvhu.update({
     where: { id: created.id },
     data: {
+      ...(canAccessDraftOrgIds.length ? { canAccessDraftOrgIds } : {}),
       ...(intermediariesOrgIds.length ? { intermediariesOrgIds } : {})
     },
     include: BsvhuForElasticInclude
@@ -97,7 +147,26 @@ const getVhuFormdata = (): Prisma.BsvhuCreateInput => ({
   destinationOperationCode: null,
 
   ecoOrganismeSiret: siretify(4),
-  ecoOrganismeName: "Eco-Organisme"
+  ecoOrganismeName: "Eco-Organisme",
+
+  brokerCompanyName: "Courtier efficace",
+  brokerCompanySiret: siretify(5),
+  brokerCompanyAddress: "15 Rue des Lilas, 33000 Lille",
+  brokerCompanyContact: "Anton Spencer",
+  brokerCompanyPhone: "06 67 78 89 91",
+  brokerCompanyMail: "a.spencer@goodbroker.com",
+  brokerRecepisseNumber: "receipt number",
+  brokerRecepisseDepartment: "33",
+  brokerRecepisseValidityLimit: "2026-11-27T00:00:00.000Z",
+  traderCompanyName: "Le Négoce QVB",
+  traderCompanySiret: siretify(6),
+  traderCompanyAddress: "32 Avenue des Azalées, 33700 Mérignac",
+  traderCompanyContact: "Benjamin Turner",
+  traderCompanyPhone: "06 68 35 64 34",
+  traderCompanyMail: "b.turner@tradingalright.com",
+  traderRecepisseNumber: "receipt of the firm",
+  traderRecepisseDepartment: "33",
+  traderRecepisseValidityLimit: "2026-11-28T00:00:00.000Z"
 });
 
 export const toIntermediaryCompany = (company: Company, contact = "toto") => ({