+++ back/src/bsvhu/permissions.ts
@@ -6,14 +6,18 @@ import { Permission, checkUserPermissions } from "../permissions";
  * Retrieves organisations allowed to read a BSVHU
  */
 function readers(bsvhu: Bsvhu): string[] {
-  return [
-    bsvhu.emitterCompanySiret,
-    bsvhu.destinationCompanySiret,
-    bsvhu.transporterCompanySiret,
-    bsvhu.transporterCompanyVatNumber,
-    bsvhu.ecoOrganismeSiret,
-    ...bsvhu.intermediariesOrgIds
-  ].filter(Boolean);
+  return bsvhu.isDraft
+    ? [...bsvhu.canAccessDraftOrgIds]
+    : [
+        bsvhu.emitterCompanySiret,
+        bsvhu.destinationCompanySiret,
+        bsvhu.transporterCompanySiret,
+        bsvhu.transporterCompanyVatNumber,
+        bsvhu.ecoOrganismeSiret,
+        bsvhu.brokerCompanySiret,
+        bsvhu.traderCompanySiret,
+        ...bsvhu.intermediariesOrgIds
+      ].filter(Boolean);
 }
 
 /**
@@ -23,12 +27,17 @@ function readers(bsvhu: Bsvhu): string[] {
  * a user is not removing his own company from the BSVHU
  */
 function contributors(bsvhu: Bsvhu, input?: BsvhuInput): string[] {
+  if (bsvhu.isDraft) {
+    return [...bsvhu.canAccessDraftOrgIds];
+  }
   const updateEmitterCompanySiret = input?.emitter?.company?.siret;
   const updateDestinationCompanySiret = input?.destination?.company?.siret;
   const updateTransporterCompanySiret = input?.transporter?.company?.siret;
   const updateTransporterCompanyVatNumber =
     input?.transporter?.company?.vatNumber;
   const updateEcoOrganismeCompanySiret = input?.ecoOrganisme?.siret;
+  const updateBrokerCompanySiret = input?.broker?.company?.siret;
+  const updateTraderCompanySiret = input?.trader?.company?.siret;
 
   const updateIntermediaries = (input?.intermediaries ?? []).flatMap(i => [
     i.siret,
@@ -59,6 +68,17 @@ function contributors(bsvhu: Bsvhu, input?: BsvhuInput): string[] {
     updateEcoOrganismeCompanySiret !== undefined
       ? updateEcoOrganismeCompanySiret
       : bsvhu.ecoOrganismeSiret;
+
+  const brokerCompanySiret =
+    updateBrokerCompanySiret !== undefined
+      ? updateBrokerCompanySiret
+      : bsvhu.brokerCompanySiret;
+
+  const traderCompanySiret =
+    updateTraderCompanySiret !== undefined
+      ? updateTraderCompanySiret
+      : bsvhu.traderCompanySiret;
+
   const intermediariesOrgIds =
     input?.intermediaries !== undefined
       ? updateIntermediaries
@@ -70,6 +90,8 @@ function contributors(bsvhu: Bsvhu, input?: BsvhuInput): string[] {
     transporterCompanySiret,
     transporterCompanyVatNumber,
     ecoOrganismeCompanySiret,
+    brokerCompanySiret,
+    traderCompanySiret,
     ...intermediariesOrgIds
   ].filter(Boolean);
 }
@@ -83,7 +105,9 @@ function creators(input: BsvhuInput) {
     input.ecoOrganisme?.siret,
     input.transporter?.company?.siret,
     input.transporter?.company?.vatNumber,
-    input.destination?.company?.siret
+    input.destination?.company?.siret,
+    input.broker?.company?.siret,
+    input.trader?.company?.siret
   ].filter(Boolean);
 }