+++ back/src/bsvhu/validation/__tests__/validation.integration.ts
@@ -37,13 +37,16 @@ describe("BSVHU validation", () => {
   let transporterCompany: Company;
   let intermediaryCompany: Company;
   let ecoOrganisme: EcoOrganisme;
+  let brokerCompany: Company;
+  let traderCompany: Company;
   beforeAll(async () => {
     const emitterCompany = await companyFactory({ companyTypes: ["PRODUCER"] });
     transporterCompany = await companyFactory({
       companyTypes: ["TRANSPORTER"]
     });
     const destinationCompany = await companyFactory({
-      companyTypes: ["WASTE_VEHICLES"]
+      companyTypes: ["WASTE_VEHICLES"],
+      wasteVehiclesTypes: ["BROYEUR", "DEMOLISSEUR"]
     });
     foreignTransporter = await companyFactory({
       companyTypes: ["TRANSPORTER"],
@@ -57,12 +60,20 @@ describe("BSVHU validation", () => {
       handle: { handleBsvhu: true },
       createAssociatedCompany: true
     });
+    brokerCompany = await companyFactory({
+      companyTypes: ["BROKER"]
+    });
+    traderCompany = await companyFactory({
+      companyTypes: ["TRADER"]
+    });
     const prismaBsvhu = await bsvhuFactory({
       opt: {
         emitterCompanySiret: emitterCompany.siret,
         transporterCompanySiret: transporterCompany.siret,
         destinationCompanySiret: destinationCompany.siret,
         ecoOrganismeSiret: ecoOrganisme.siret,
+        brokerCompanySiret: brokerCompany.siret,
+        traderCompanySiret: traderCompany.siret,
         intermediaries: {
           create: [toIntermediaryCompany(intermediaryCompany)]
         }
@@ -353,9 +364,7 @@ describe("BSVHU validation", () => {
         expect((err as ZodError).issues).toEqual([
           expect.objectContaining({
             message:
-              `L'installation de destination avec le SIRET \"${company.siret}\" n'est pas inscrite` +
-              " sur Trackdéchets en tant qu'installation de traitement de VHU. Cette installation ne peut donc pas" +
-              " être visée sur le bordereau. Veuillez vous rapprocher de l'administrateur de cette installation pour qu'il modifie le profil de l'établissement depuis l'interface Trackdéchets dans Mes établissements"
+              "Cet établissement n'a pas le profil Installation de traitement de VHU."
           })
         ]);
       }
@@ -608,7 +617,9 @@ describe("BSVHU validation", () => {
         [bsvhu.transporterCompanySiret!]: searchResult("transporteur"),
         [bsvhu.destinationCompanySiret!]: searchResult("destinataire"),
         [intermediaryCompany.siret!]: searchResult("intermédiaire"),
-        [ecoOrganisme.siret!]: searchResult("ecoOrganisme")
+        [ecoOrganisme.siret!]: searchResult("ecoOrganisme"),
+        [brokerCompany.siret!]: searchResult("broker"),
+        [traderCompany.siret!]: searchResult("trader")
       };
       (searchCompany as jest.Mock).mockImplementation((clue: string) => {
         return Promise.resolve(searchResults[clue]);
@@ -644,6 +655,12 @@ describe("BSVHU validation", () => {
       );
       expect(sirenified.ecoOrganismeName).toEqual(
         searchResults[ecoOrganisme.siret!].name
+      );
+      expect(sirenified.brokerCompanyName).toEqual(
+        searchResults[brokerCompany.siret!].name
+      );
+      expect(sirenified.traderCompanyName).toEqual(
+        searchResults[traderCompany.siret!].name
       );
     });
     it("should not overwrite `name` and `address` based on SIRENE data for sealed fields", async () => {